requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

fastapi>=0.95.0
uvicorn[standard]>=0.20.0
